\name{sem1R-package}
\alias{sem1R-package}
\docType{package}
\title{Package finds interesting, non-trivial and hidden patterns in omics data.}
\description{
	For inducing semantic rules we require to provide a 2-dimensional binary matrix, ontologies in OBO format, and description of rows and columns which associate rows/columns with the corresponding row/column ontology.
}
\details{
	The package consists of only one class sem1R that stores all parameters and results.
}
\author{
Frantisek Malinka, frantisek.malinka@gmail.com.

Maintainer: Frantisek Malinka <frantisek.malinka@gmail.com>
}
\references{
  Malinka, F., Zelezny, F., Klema, J. Concept rule learning with and ontology-based refinement operator inducing semantics rule from omics data. Unpublished. \cr\cr
  For more information see Github \url{http://www.github.com/fmalinka/sem1r}
}
